requirements.txt
numpy>=1.21.0
pandas>=1.3.0
PyYAML>=5.4.1
requests>=2.26.0
pytz>=2021.1
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
discord-webhook>=0.14.0
python-dateutil>=2.8.2
schedule>=1.2.1 